requirements.txt
speedtest-cli
mysql-connector-python
